Support both static and dynamic linking
