+++ .github/workflows/pythonpackage.yml
@@ -27,7 +27,6 @@ jobs:
         sudo apt install -yqq g++8 cmake make libboost-dev libboost-date-time-dev
     - name: Patch abieos build files
       run: |
-        sed -i 's/-fsanitize=address,undefined//' external/abieos/CMakeLists.txt
         sed -i 's/STATIC src\/abi.cpp/STATIC src\/abieos.cpp src\/abi.cpp/' external/abieos/CMakeLists.txt
     - name: Build abieos
       run: |
@@ -35,7 +34,7 @@ jobs:
         mkdir build
         cd build
         CC=gcc-8 CXX=g++-8 cmake ..
-        make
+        make abieos
     - uses: actions/upload-artifact@v1
       with:
         name: abieos
@@ -62,7 +61,7 @@ jobs:
         python-version: ${{ matrix.python-version }}
     - name: Build library
       run: |
-        python setup.py build
+        python setup.py build --static
         cp build/lib.*/abieos/_private* abieos
     - name: Install dev dependencies
       run: |
